ExtensionManager methods.

+++ components/repositories_manager.py
@@ -4,7 +4,7 @@ import re
 import os
 import shutil
 
-from model import Extractor, Inspector
+from model import Extractor, Inspector, OperationError
 
 
 def camel_case_to_underscore(name):
@@ -41,10 +41,12 @@ class RepositoriesManager(object):
         :rtype : None
         """
         dest_path = os.path.join(self.repositories_dir_name, repo_name)
-        try:
-            shutil.copyfile(file_path, dest_path)
-        except IOError:
-            raise
+        file_name = os.path.basename(file_path)
+
+        if os.path.exists(os.path.join(dest_path, file_name)):
+            raise OperationError("The file '{0}' already exists on the repository '{1}'".format(file_name, repo_name))
+
+        shutil.copy(file_path, dest_path)
 
     def remove_file(self, repo_name, file_name):
         """
@@ -53,10 +55,12 @@ class RepositoriesManager(object):
         :rtype : None
         """
         target_path = os.path.join(self.repositories_dir_name, repo_name, file_name)
-        try:
-            os.remove(target_path)
-        except OSError:
-            raise
+
+        if not os.path.exists(target_path):
+            return
+
+        os.remove(target_path)
+
 
     def get_extractor_instance(self, name):
         """
